The key changes are:
In the Serve Shiny app step, we're using Rscript -e "source('app.R')" to source the Shiny app code from the app.R file in the top-level folder.
We're then using the app variable (which is defined in the app.R file) to run the Shiny app using runApp(app, port = 80, host = '0.0.0.0').